backend/pyproject.toml
[tool.poetry]
name = "tunescore-backend"
version = "0.1.0"
description = "AI-powered music industry intelligence platform"
authors = ["dwood <dwood@tunescore.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.10.0"}
pydantic-settings = "^2.7.0"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.2.0"
httpx = "^0.28.0"
requests-cache = "^1.2.1"
aiofiles = "^24.1.0"
python-dotenv = "^1.0.1"
structlog = "^24.4.0"
tenacity = "^9.0.0"

# Music/Audio Processing
librosa = "^0.10.0"
soundfile = "^0.12.1"
pydub = "^0.25.1"
numpy = "^1.26.0"
scipy = "^1.11.0"

# Advanced ML-based Music Analysis (Hugging Face models)
transformers = "^4.35.0"  # Pre-trained genre & instrument models
torch = "^2.1.0"  # PyTorch backend
torchaudio = "^2.1.0"  # Audio processing

# Advanced audio analysis
madmom = "^0.16.1"  # Beat tracking and onset detection
demucs = "^4.0.1"  # Source separation

# NLP/Embeddings
vaderSentiment = "^3.3.2"
sentence-transformers = "^2.2.0"
spacy = "^3.7.0"  # NER and linguistic features
langdetect = "^1.0.9"  # Language detection
deep-translator = "^1.11.4"  # Translation

# Industry Pulse - Web Scraping & Data Aggregation
beautifulsoup4 = "^4.12.0"
feedparser = "^6.0.11"
lxml = "^5.1.0"
html5lib = "^1.1"
apscheduler = "^3.10.4"

# Music Industry Integrations
spotipy = "^2.23.0"
google-api-python-client = "^2.100.0"
musicbrainzngs = "^0.7.1"  # MusicBrainz API for credits

# ML Models
xgboost = "^2.0.0"  # Gradient boosting for predictions
scikit-learn = "^1.3.0"  # Already included via dependencies, making explicit

# AI Services (Optional)
anthropic = "^0.40.0"
openai = "^1.58.0"
openai-whisper = "^20250625"
pyloudnorm = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
httpx = "^0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
target-version = "py312"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused-function-args
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "T20", # flake8-print
    "PIE", # flake8-pie
    "RET", # flake8-return
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from - intentionally re-raising as HTTPException in API routes
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "T20",   # print statements (allowed in debug scripts)
]
unfixable = ["B"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"app/tests/*" = ["ARG001", "S101"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "anthropic.*",
    "openai.*",
    "librosa.*",
    "soundfile.*",
    "pydub.*",
    "vaderSentiment.*",
    "sentence_transformers.*",
    "spotipy.*",
    "madmom.*",
    "demucs.*",
    "essentia.*",
    "langdetect.*",
    "deep_translator.*",
    "musicbrainzngs.*",
    "xgboost.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["app/tests"]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
